+++ ddhcp.h
@@ -8,12 +8,12 @@
 int ddhcp_block_init(ddhcp_config* config);
 void ddhcp_block_free(ddhcp_config* config);
 
-void ddhcp_block_process(uint8_t* buffer, int len, struct sockaddr_in6 sender, ddhcp_config* config);
+void ddhcp_block_process(uint8_t* buffer, ssize_t len, struct sockaddr_in6 sender, ddhcp_config* config);
 
 void ddhcp_block_process_claims(struct ddhcp_mcast_packet* packet, ddhcp_config* config);
 void ddhcp_block_process_inquire(struct ddhcp_mcast_packet* packet, ddhcp_config* config);
 
-void ddhcp_dhcp_process(uint8_t* buffer, int len, struct sockaddr_in6 sender, ddhcp_config* config);
+void ddhcp_dhcp_process(uint8_t* buffer, ssize_t len, struct sockaddr_in6 sender, ddhcp_config* config);
 void ddhcp_dhcp_renewlease(struct ddhcp_mcast_packet* packet, ddhcp_config* config);
 void ddhcp_dhcp_leaseack(struct ddhcp_mcast_packet* packet, ddhcp_config* config);
 void ddhcp_dhcp_leasenak(struct ddhcp_mcast_packet* packet, ddhcp_config* config);
